+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-04-18
+
+### Added
+
+- Add Client#classifications to predict the most likely labels based on examples or a file.
+
+### Fixed
+
+- Fixed Files#upload which was previously broken by the validation code!
+
 ## [1.2.2] - 2021-04-18
 
 ### Changed